Lower SoC poll intervall if not charging but battery almost full
If evcc is set to 100% charging, it will keep charging enabled, as long
as charging mode allows and not stop at reported 100% but when car stops
charging.
So in this case, evcc will remain in charging_enabled, but not charging,
and in this case it's not sensile to wake up the HV system every minute.
In this corner case, the poll interval is 1h, so that _if_ the car did
not wake up on the charging_enable poll, it will be corrected over time,
but not wake up the system too often.